Changing how we pass errors thru form components
having a single source of errors was confusing and overloaded. Now we
always pass down ALL errors via an `errors` prop and only errors
specific to the component in question via `fieldErrors`. Each component
expects `errors`to be an array and `fiedlErrors` to be specific to be
specific to what the component needs.