Changing how trees are constructed
Before, trees accept an object with ref and refs. This commit removes
the object and renames refs -> value and refs -> children.

The impetus to rename the parameters stemmed from making them actually
match what the tree expects. Having them as ref and refs tightly
coupled the tree to react.

The impetus for removing the object is just to clean up noise.

I am open to change in any way