🔧 fix issues with package.json
Honestly, emacs kills me a little every time I leave in a "u" or
a "p" in the source code.

This error in particular was an extra "p" in one unmocked module
path for the Jest configuration, which throws really obscure errors.

Thankfully this pointed me in the right direction:

https://github.com/fedosejev/react-essentials/issues/19#issuecomment-169001776

But holy 💩. Worst debug session ever, emacs is the worst.